requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
pandas
google-generativeai
python-dotenv
sqlalchemy
firecrawl-py
